Refactored pie chart
